Add support of creating a new file from a link next to current one
In Obsidian it's possible to configure the location of a new file:
* Vault root
* next to the current file
* Specific one (the inbox).

In obsidian.el it was only the third, the inbox directory.

With the current change the second option is also available,
next to the current file, in the same directory, if
the package :config contains:
 (setq obsidian-wiki-link-create-file-in-inbox nil)
Here the `t` is the default value, to store in the inbox.

Fixes issue #74.