obsidian--match-files finds all files with given basename & more (#73)
When issue #64 was opened, obsidian--match-files was lazy,
only the suffix was checked, so it found for 'aa.md' both
'aa.md' and 'baa.md'. Its solution broke the relative file name
usage (e.g. [[aa.md]] was "found" in the inbox).

To cover both use case the match needs to be either exact
(relative to the vault root), or partial, where the searched
file name prefixed with '/' is searched with s-ends-with-p.

Fixes issue #72.