Add obsidian-apply-template interactive command
Previously the daily notes command used `obsidian-apply-template`,
but it is internal, so it gets the `obsidian--apply-template`.

For the interactive version, the original `obsidian-apply-tempalte`
is used. It list the template files from the templates directory
without extension, like `Daily Note Template`.
If an existing file name is chosen, it is applied, inserted
to the beginning of the current buffer.